requirements.txt
# Core async framework
asyncio-mqtt==0.16.1

# MongoDB async driver
motor==3.3.2

# Polymarket CLOB API client
py-clob-client==0.28.0

# WebSocket for Binance
websockets==12.0

# Configuration management
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv==1.0.0

# Telegram bot
python-telegram-bot==20.7

# CLI dashboard
rich==13.7.0

# HTTP client
aiohttp==3.9.1
httpx~=0.25.2

# Utilities
python-dateutil==2.8.2

# Testing (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
